Change behavior of vendor option
Setting vendor to true lints copied vendor file regardless
of copy option. Compiled vendor files are never linted.

This is to have the same behavior as jshint, jshint.vendor
documentation:

jshint.vendor
When this property is set to true, Mimosa will execute jshint
on the successful copy of a vendor JavaScript asset.